refactor: 사용하지 않는 print문 제거
